Record Counts by View
A coworker had asked me for a quick way to get row counts for all views
in the current database.  Unlike tables, getting row counts for views
must be done by selecting the view :(.
This script will generate the sql statements for all views in the
current database and enumerate the view with the number of rows.